add vm config to starcoin-types
