+++ types/src/vm/mod.rs
@@ -1,0 +1,4 @@
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod config;
