+++ vm/vm-runtime/src/move_vm_ext/vm.rs
@@ -5,31 +5,24 @@ use crate::move_vm_ext::session::SessionId;
 use crate::move_vm_ext::warm_vm_cache::WarmVmCache;
 use crate::move_vm_ext::{SessionExt, StarcoinMoveResolver};
 use crate::natives;
-use move_binary_format::deserializer::DeserializerConfig;
-use move_binary_format::{
-    errors::VMResult,
-    file_format_common,
-    file_format_common::{IDENTIFIER_SIZE_MAX, LEGACY_IDENTIFIER_SIZE_MAX},
-};
-use move_bytecode_verifier::VerifierConfig;
-use move_vm_runtime::config::VMConfig;
+use move_binary_format::errors::VMResult;
 use move_vm_runtime::move_vm::MoveVM;
 use move_vm_runtime::native_extensions::NativeContextExtensions;
 use move_vm_types::loaded_data::runtime_types::TypeBuilder;
+use starcoin_framework::natives::{
+    aggregator_natives::NativeAggregatorContext, event::NativeEventContext,
+    object::NativeObjectContext,
+};
 use starcoin_gas_algebra::DynamicExpression;
 use starcoin_gas_schedule::{MiscGasParameters, NativeGasParameters};
 use starcoin_native_interface::SafeNativeBuilder;
 use starcoin_table_natives::NativeTableContext;
+use starcoin_types::vm::config::starcoin_prod_vm_config;
 use starcoin_vm_types::{
     errors::PartialVMResult,
-    on_chain_config::{
-        FeatureFlag, Features, TimedFeatureFlag, TimedFeatures, TimedFeaturesBuilder,
-    },
+    on_chain_config::{Features, TimedFeatureFlag, TimedFeatures, TimedFeaturesBuilder},
 };
 
-use starcoin_framework::natives::aggregator_natives::NativeAggregatorContext;
-use starcoin_framework::natives::event::NativeEventContext;
-use starcoin_framework::natives::object::NativeObjectContext;
 use std::ops::Deref;
 use std::sync::Arc;
 
@@ -37,32 +30,6 @@ pub struct MoveVmExt {
     inner: MoveVM,
     _chain_id: u8,
     features: Arc<Features>,
-}
-
-pub fn get_max_binary_format_version(
-    features: &Features,
-    gas_feature_version_opt: Option<u64>,
-) -> u32 {
-    // For historical reasons, we support still < gas version 5, but if a new caller don't specify
-    // the gas version, we default to 5, which was introduced in late '22.
-    let gas_feature_version = gas_feature_version_opt.unwrap_or(5);
-    if gas_feature_version < 5 {
-        file_format_common::VERSION_5
-    } else if features.is_enabled(FeatureFlag::VM_BINARY_FORMAT_V7) {
-        file_format_common::VERSION_7
-    } else if features.is_enabled(FeatureFlag::VM_BINARY_FORMAT_V6) {
-        file_format_common::VERSION_6
-    } else {
-        file_format_common::VERSION_5
-    }
-}
-
-pub fn get_max_identifier_size(features: &Features) -> u64 {
-    if features.is_enabled(FeatureFlag::LIMIT_MAX_IDENTIFIER_LENGTH) {
-        IDENTIFIER_SIZE_MAX
-    } else {
-        LEGACY_IDENTIFIER_SIZE_MAX
-    }
 }
 
 impl MoveVmExt {
@@ -79,29 +46,11 @@ impl MoveVmExt {
     where
         F: Fn(DynamicExpression) + Send + Sync + 'static,
     {
-        // Note: binary format v6 adds a few new integer types and their corresponding instructions.
-        //       Therefore it depends on a new version of the gas schedule and cannot be allowed if
-        //       the gas schedule hasn't been updated yet.
-        let max_binary_format_version =
-            get_max_binary_format_version(&features, Some(gas_feature_version));
-
-        let max_identifier_size = get_max_identifier_size(&features);
+        let vm_config = starcoin_prod_vm_config(&features, &timed_features, TypeBuilder::Legacy);
 
         let _enable_invariant_violation_check_in_swap_loc =
             !timed_features.is_enabled(TimedFeatureFlag::DisableInvariantViolationCheckInSwapLoc);
         let _type_size_limit = true;
-
-        let verifier_config = verifier_config(&features, &timed_features);
-
-        let mut type_max_cost = 0;
-        let mut type_base_cost = 0;
-        let mut type_byte_cost = 0;
-        if timed_features.is_enabled(TimedFeatureFlag::LimitTypeTagSize) {
-            // 5000 limits type tag total size < 5000 bytes and < 50 nodes
-            type_max_cost = 5000;
-            type_base_cost = 100;
-            type_byte_cost = 1;
-        }
 
         let mut builder = SafeNativeBuilder::new(
             gas_feature_version,
@@ -115,23 +64,7 @@ impl MoveVmExt {
             builder.set_gas_hook(hook);
         }
         Ok(Self {
-            inner: WarmVmCache::get_warm_vm(
-                builder,
-                VMConfig {
-                    verifier_config,
-                    deserializer_config: DeserializerConfig::new(max_binary_format_version, max_identifier_size),
-                    paranoid_type_checks: /*crate::StarcoinVM::get_paranoid_checks() */ false,
-                    max_value_nest_depth: Some(128),
-                    type_max_cost,
-                    type_base_cost,
-                    type_byte_cost,
-                    // todo: support aggregator_v2_type_tagging, set false as default now.
-                    aggregator_v2_type_tagging: false,
-                    check_invariant_in_swap_loc: false,
-                    ty_builder: TypeBuilder::Legacy,
-                },
-                resolver,
-            )?,
+            inner: WarmVmCache::get_warm_vm(builder, vm_config, resolver)?,
             _chain_id,
             features: Arc::new(features),
         })
@@ -231,29 +164,5 @@ impl Deref for MoveVmExt {
 
     fn deref(&self) -> &Self::Target {
         &self.inner
-    }
-}
-
-pub fn verifier_config(features: &Features, _timed_features: &TimedFeatures) -> VerifierConfig {
-    VerifierConfig {
-        max_loop_depth: Some(5),
-        max_generic_instantiation_length: Some(32),
-        max_function_parameters: Some(128),
-        max_basic_blocks: Some(1024),
-        max_value_stack_size: 1024,
-        max_type_nodes: Some(256),
-        max_dependency_depth: Some(256),
-        max_push_size: Some(10000),
-        max_struct_definitions: None,
-        max_fields_in_struct: None,
-        max_function_definitions: None,
-        max_back_edges_per_function: None,
-        max_back_edges_per_module: None,
-        max_basic_blocks_in_script: None,
-        max_per_fun_meter_units: Some(1000 * 80000),
-        max_per_mod_meter_units: Some(1000 * 80000),
-        use_signature_checker_v2: features.is_enabled(FeatureFlag::SIGNATURE_CHECKER_V2),
-        sig_checker_v2_fix_script_ty_param_count: features
-            .is_enabled(FeatureFlag::SIGNATURE_CHECKER_V2_SCRIPT_FIX),
     }
 }
